Refactor: move withCurrentScope from Scope.Monad to new Scope.State
