+++ Agda.cabal
@@ -683,6 +683,7 @@ library
       Agda.Syntax.Scope.Base
       Agda.Syntax.Scope.Monad
       Agda.Syntax.Scope.Operator
+      Agda.Syntax.Scope.State
       Agda.Syntax.TopLevelModuleName
       Agda.Syntax.TopLevelModuleName.Boot
       Agda.Syntax.Translation.AbstractToConcrete
